GCP_Agent_Starter_Pack/winsights_agent/requirements.txt
absl-py
google-adk
google-cloud-aiplatform
python-dotenv
pandas
numpy
matplotlib
plotly
